JNM: Extract JavaNativeHeaderConverter
